feat: now all qutip's operators gates, states can be used by using the jaxify function

+++ feedback_grape/utils/gates.py
@@ -12,7 +12,10 @@ def cnot():
     """
     Controlled NOT gate.
     """
-    return jnp.array([[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]])
+    return jnp.array(
+        [[1, 0, 0, 0], [0, 1, 0, 0], [0, 0, 0, 1], [0, 0, 1, 0]],
+        dtype=jnp.complex128,
+    )
 
 
 # TODO: Check for the hadamard definition from qutip for n qubits
@@ -20,4 +23,4 @@ def hadamard():
     """
     Hadamard transform operator.
     """
-    return jnp.array([[1, 1], [1, -1]]) / jnp.sqrt(2)
+    return jnp.array([[1, 1], [1, -1]], dtype=jnp.complex128) / jnp.sqrt(2)
